requirements.txt
mcp>=1.2.0
psutil
streamlit
# RAG Dependencies
langchain
langchain-community
langchain-ollama
langchain-chroma
chromadb
supabase
pypdf
tiktoken
python-dotenv
# Guardrails
pydantic
